Add files via upload
A simple script to listen to and record the result of the [Jito REST API](https://jito-labs.gitbook.io/mev/searcher-resources/bundles#tip-guidelines) that describes the percentile tip values required to land a transaction.

In developing this service, I realized that a) the websocket is unstable - in setting an initial listening preference to once every 3s, I got rate-limited immediately to once every 30s. Pinging the REST API every 3s appears to be stable and b) the data only update once every minute anyways.

Regardless, helpful. I may try to publish a txn fee monitor based on this data.